requirement.txt
pyqt5==5.13.0
pyqt5-tools
matplotlib==3.0.2
python-telegram-bot==12.3.0
requests==2.22.0
urllib3==1.25.3
openpyxl==3.0.3